+++ aztk/models/__init__.py
@@ -1,1 +1,2 @@
+from .toolkit import Toolkit, TOOLKIT_MAP
 from .models import *