+++ aztk_cli/config.py
@@ -10,6 +10,7 @@ from aztk.spark.models import (
     ClusterConfiguration,
     UserConfiguration,
 )
+from aztk.models import Toolkit
 from aztk.models.plugins.internal import PluginReference
 
 def load_aztk_secrets() -> SecretsConfiguration:
@@ -185,8 +186,8 @@ def cluster_config_from_dict(config: dict):
                     mount_path=file_share['mount_path'],
                 ))
 
-    if config.get('docker_repo') is not None:
-        output.docker_repo = config['docker_repo']
+    if config.get('toolkit') is not None:
+        output.toolkit = Toolkit.from_dict(config['toolkit'])
 
     if config.get('plugins') not in [[None], None]:
         output.plugins = []
@@ -299,7 +300,7 @@ class JobConfig():
         self.custom_scripts = None
         self.spark_configuration = None
         self.vm_size = None
-        self.docker_repo = None
+        self.toolkit = None
         self.max_dedicated_nodes = 0
         self.max_low_pri_nodes = 0
         self.spark_defaults_conf = None
@@ -317,7 +318,7 @@ class JobConfig():
         cluster_configuration = config.get('cluster_configuration')
         if cluster_configuration:
             self.vm_size = cluster_configuration.get('vm_size')
-            self.docker_repo = cluster_configuration.get('docker_repo')
+            self.toolkit = Toolkit.from_dict(cluster_configuration.get('toolkit'))
             if cluster_configuration.get('size') is not None:
                 self.max_dedicated_nodes = cluster_configuration.get('size')
             if cluster_configuration.get('size_low_pri') is not None: