+++ aztk_cli/spark/endpoints/cluster/cluster_create.py
@@ -49,8 +49,7 @@ def execute(args: typing.NamedTuple):
         user_configuration=UserConfiguration(
             username=args.username,
             password=args.password,
-        ),
-        docker_repo=args.docker_repo))
+        )))
     wait = wait if args.wait is None else args.wait
 
     user_configuration = cluster_conf.user_configuration